First Content commit
Homestead is a new child theme for Hello Elementor by David Battiston.
It was built to fix problems related to the Uusi theme.

Main Advantages:
No longer includes main theme, Hello Elementor is a dependancy that is
maintained by it's own team. Bootstrap no longer needs to be manually
updated.

- David Battiston